rtemslwip/xil_shims: Perform flush with invalidate
Xilinx wrote their A53 HAL with the assumption that the CPU did not
support cache invalidation without a flush, so the flush and
invalidation functions were combined and all range invalidations are
promoted to flush/invalidate. The implementation written for lwIP was
written to the original intent of the function and thus was not flushing
in some cases when it needed to. This resolves that issue which prevents
DMA transmit errors in some cases.